# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/pets": {
      "post": {
        "operationId": "create",
        "summary": "Create a new pet",
        "description": "Create a new pet for the authenticated user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pet created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid pet data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all pets",
        "description": "Retrieve all pets with optional filtering and pagination",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "species",
            "required": false,
            "in": "query",
            "description": "Filter by pet species",
            "schema": {
              "enum": [
                "dog",
                "cat",
                "bird",
                "rabbit",
                "hamster",
                "fish",
                "reptile",
                "horse",
                "other"
              ],
              "type": "string"
            }
          },
          {
            "name": "gender",
            "required": false,
            "in": "query",
            "description": "Filter by pet gender",
            "schema": {
              "enum": [
                "male",
                "female",
                "unknown"
              ],
              "type": "string"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "description": "Filter by pet size",
            "schema": {
              "enum": [
                "tiny",
                "small",
                "medium",
                "large",
                "giant"
              ],
              "type": "string"
            }
          },
          {
            "name": "is_spayed_neutered",
            "required": false,
            "in": "query",
            "description": "Filter by spayed/neutered status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "is_vaccinated",
            "required": false,
            "in": "query",
            "description": "Filter by vaccination status",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search in name, breed, or color",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner_id",
            "required": false,
            "in": "query",
            "description": "Filter by owner ID (admin only)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pets retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pets/my-pets": {
      "get": {
        "operationId": "findMyPets",
        "summary": "Get current user pets",
        "description": "Retrieve all pets owned by the authenticated user",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User pets retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pets/stats": {
      "get": {
        "operationId": "getPetStats",
        "summary": "Get pet statistics",
        "description": "Retrieve comprehensive statistics about all pets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Pet statistics retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pets/species/{species}": {
      "get": {
        "operationId": "getPetsBySpecies",
        "summary": "Get pets by species",
        "description": "Retrieve all pets of a specific species",
        "parameters": [
          {
            "name": "species",
            "required": true,
            "in": "path",
            "description": "Pet species to filter by",
            "schema": {
              "enum": [
                "dog",
                "cat",
                "bird",
                "rabbit",
                "hamster",
                "fish",
                "reptile",
                "horse",
                "other"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pets by species retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pets/needing-vaccination": {
      "get": {
        "operationId": "getPetsNeedingVaccination",
        "summary": "Get pets needing vaccination",
        "description": "Retrieve all pets that need vaccination",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Pets needing vaccination retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pets/needing-spay-neuter": {
      "get": {
        "operationId": "getPetsNeedingSpayNeuter",
        "summary": "Get pets needing spay/neuter",
        "description": "Retrieve all pets that need spay/neuter surgery",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Pets needing spay/neuter retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pets/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get pet by ID",
        "description": "Retrieve a specific pet by its ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pet ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "update",
        "summary": "Update pet",
        "description": "Update a specific pet by its ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pet ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pet updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pet"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid update data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete pet",
        "description": "Soft delete a specific pet by its ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pet ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pets/{id}/hard": {
      "delete": {
        "operationId": "hardRemove",
        "summary": "Hard delete pet",
        "description": "Permanently delete a specific pet by its ID (admin only)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Pet ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet permanently deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - admin access required"
          },
          "404": {
            "description": "Pet not found"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/pets/validate": {
      "post": {
        "operationId": "validatePetData",
        "summary": "Validate pet data",
        "description": "Validate pet data without creating the pet",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pet data validation completed"
          },
          "400": {
            "description": "Bad request - invalid pet data"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "pets"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/appointments/pet/{petId}": {
      "get": {
        "operationId": "findByPet",
        "summary": "Get appointments by pet",
        "description": "Retrieve all appointments for a specific pet",
        "parameters": [
          {
            "name": "petId",
            "required": true,
            "in": "path",
            "description": "Pet ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pet appointments retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "appointments"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Clinic Management API",
    "description": "AI-Powered Pet Clinic Management Platform Backend API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreatePetDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the pet",
            "example": "Buddy"
          },
          "species": {
            "type": "string",
            "description": "Species of the pet",
            "enum": [
              "dog",
              "cat",
              "bird",
              "rabbit",
              "hamster",
              "fish",
              "reptile",
              "horse",
              "other"
            ],
            "example": "dog"
          },
          "breed": {
            "type": "string",
            "description": "Breed of the pet",
            "example": "Golden Retriever"
          },
          "gender": {
            "type": "string",
            "description": "Gender of the pet",
            "enum": [
              "male",
              "female",
              "unknown"
            ],
            "example": "male"
          },
          "date_of_birth": {
            "type": "string",
            "description": "Date of birth of the pet",
            "example": "2020-03-15"
          },
          "weight": {
            "type": "number",
            "description": "Weight of the pet in pounds",
            "example": 45.5
          },
          "size": {
            "type": "string",
            "description": "Size category of the pet",
            "enum": [
              "tiny",
              "small",
              "medium",
              "large",
              "giant"
            ],
            "example": "medium"
          },
          "color": {
            "type": "string",
            "description": "Color/markings of the pet",
            "example": "Golden"
          },
          "microchip_number": {
            "type": "string",
            "description": "Microchip number if available",
            "example": "123456789012345"
          },
          "is_spayed_neutered": {
            "type": "boolean",
            "description": "Whether the pet is spayed/neutered",
            "example": true
          },
          "is_vaccinated": {
            "type": "boolean",
            "description": "Whether the pet is vaccinated",
            "example": true
          },
          "medical_history": {
            "type": "string",
            "description": "Medical history and notes",
            "example": "No known health issues"
          },
          "behavioral_notes": {
            "type": "string",
            "description": "Behavioral notes",
            "example": "Friendly with other dogs"
          },
          "dietary_requirements": {
            "type": "string",
            "description": "Special dietary requirements",
            "example": "Grain-free diet"
          },
          "allergies": {
            "description": "Allergies and sensitivities",
            "example": [
              "Peanuts",
              "Dairy"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "medications": {
            "description": "Current medications",
            "example": [
              "Heartgard",
              "Flea treatment"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "emergency_contact": {
            "type": "string",
            "description": "Emergency contact information",
            "example": "John Doe"
          },
          "emergency_phone": {
            "type": "string",
            "description": "Emergency phone number",
            "example": "+1234567890"
          },
          "photo_url": {
            "type": "string",
            "description": "Profile photo URL",
            "example": "https://example.com/pet-photo.jpg"
          }
        },
        "required": [
          "name",
          "species"
        ]
      },
      "Pet": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for the pet"
          },
          "name": {
            "type": "string",
            "description": "Name of the pet"
          },
          "species": {
            "type": "string",
            "description": "Species of the pet",
            "enum": [
              "dog",
              "cat",
              "bird",
              "rabbit",
              "hamster",
              "fish",
              "reptile",
              "horse",
              "other"
            ]
          },
          "breed": {
            "type": "string",
            "description": "Breed of the pet"
          },
          "gender": {
            "type": "string",
            "description": "Gender of the pet",
            "enum": [
              "male",
              "female",
              "unknown"
            ]
          },
          "date_of_birth": {
            "format": "date-time",
            "type": "string",
            "description": "Date of birth of the pet"
          },
          "weight": {
            "type": "number",
            "description": "Weight of the pet in pounds"
          },
          "size": {
            "type": "string",
            "description": "Size category of the pet",
            "enum": [
              "tiny",
              "small",
              "medium",
              "large",
              "giant"
            ]
          },
          "color": {
            "type": "string",
            "description": "Color/markings of the pet"
          },
          "microchip_number": {
            "type": "string",
            "description": "Microchip number if available"
          },
          "is_spayed_neutered": {
            "type": "boolean",
            "description": "Whether the pet is spayed/neutered"
          },
          "is_vaccinated": {
            "type": "boolean",
            "description": "Whether the pet is vaccinated"
          },
          "medical_history": {
            "type": "string",
            "description": "Medical history and notes"
          },
          "behavioral_notes": {
            "type": "string",
            "description": "Behavioral notes"
          },
          "dietary_requirements": {
            "type": "string",
            "description": "Special dietary requirements"
          },
          "allergies": {
            "description": "Allergies and sensitivities",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "medications": {
            "description": "Current medications",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "emergency_contact": {
            "type": "string",
            "description": "Emergency contact information"
          },
          "emergency_phone": {
            "type": "string",
            "description": "Emergency phone number"
          },
          "photo_url": {
            "type": "string",
            "description": "Profile photo URL"
          },
          "is_active": {
            "type": "boolean",
            "description": "Whether the pet is active"
          },
          "created_at": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the pet was created"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string",
            "description": "Date when the pet was last updated"
          },
          "owner_id": {
            "type": "string",
            "description": "ID of the pet owner"
          }
        },
        "required": [
          "id",
          "name",
          "species",
          "breed",
          "gender",
          "date_of_birth",
          "weight",
          "size",
          "color",
          "microchip_number",
          "is_spayed_neutered",
          "is_vaccinated",
          "medical_history",
          "behavioral_notes",
          "dietary_requirements",
          "allergies",
          "medications",
          "emergency_contact",
          "emergency_phone",
          "photo_url",
          "is_active",
          "created_at",
          "updated_at",
          "owner_id"
        ]
      },
...
    }
  }
}